@@ -1,5 +1,9 @@
 package fr.marieteam.pdf.marieteampdf;
 
+import java.io.File;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+
 /**
  * Classe Launcher qui sert de point d'entrée pour le programme packagé
  * Cette classe est nécessaire pour contourner certaines limitations de JavaFX
@@ -11,23 +15,26 @@ public class Launcher {
      * Nécessaire pour le packaging en exécutable natif
      */
     public static void main(String[] args) {
-        // On vérifie que le dossier images est accessible
-        java.io.File imagesDir = new java.io.File("images");
+        // Définir le chemin des images de manière relative
+        String userDir = System.getProperty("user.dir");
+        Path imagesPath = Paths.get(userDir, "images");
+        File imagesDir = imagesPath.toFile();
+
         if (!imagesDir.exists()) {
-            System.out.println("Dossier images non trouvé à : " + imagesDir.getAbsolutePath());
-            // Essayer de créer le dossier si besoin
-            imagesDir.mkdirs();
-            // Afficher le répertoire de travail actuel pour le débogage
-            System.out.println("Répertoire de travail actuel : " + System.getProperty("user.dir"));
+            System.out.println("Création du dossier images à : " + imagesDir.getAbsolutePath());
+            if (!imagesDir.mkdirs()) {
+                System.err.println("Erreur : Impossible de créer le dossier images");
+            }
         } else {
             System.out.println("Dossier images trouvé à : " + imagesDir.getAbsolutePath());
-            // Lister les fichiers du dossier images
-            java.io.File[] files = imagesDir.listFiles();
-            if (files != null) {
+            File[] files = imagesDir.listFiles();
+            if (files != null && files.length > 0) {
                 System.out.println("Fichiers dans le dossier images :");
-                for (java.io.File file : files) {
+                for (File file : files) {
                     System.out.println(" - " + file.getName());
                 }
+            } else {
+                System.out.println("Le dossier images est vide");
             }
         }
         
